move array impl to new grids module
