+++ gridly/Cargo.toml
@@ -7,11 +7,7 @@ readme = "README.md"
 license = "MPL-2.0"
 description = "A library for managing fixed-size 2D spaces"
 repository = "https://github.com/Lucretiel/gridly-rs"
-categories = [
-	"no-std",
-	"data-structures",
-	"games"
-]
+categories = ["no-std", "data-structures", "games"]
 
 [badges]
 travis-ci = { repository = "Lucretiel/gridly-rs" }
@@ -19,6 +15,3 @@ maintenance = { status = "actively-developed" }
 
 [dev-dependencies]
 cool_asserts = "1.0.0"
-
-[features]
-array_impls = []
